backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
tensorflow
numpy
python-multipart
transformers
torch
pandas
scikit-learn
llama-index-core
llama-index-readers-file
requests
rq
redis
boto3
botocore
pydub
elevenlabs
prometheus-client
sentry-sdk
pytest
pytest-cov
pytest-asyncio
httpx
tf-keras
python-dotenv